Rename: 초대 페이지 폴더 구조 수정
