+++ src/components/feature/index.ts
@@ -1,1 +1,2 @@
 export * from './group'
+export * from './invite'
