+++ src/components/feature/invite/index.ts
@@ -1,0 +1,1 @@
+export * from './invite-card'
