+++ src/components/feature/invite/invite-card/index.ts
@@ -1,0 +1,1 @@
+export { InviteCard, InviteCardSkeleton } from './InviteCard'
