+++ src/pages/index.ts
@@ -4,3 +4,4 @@ export * from './point'
 export * from './auth'
 export * from './profile'
 export * from './main'
+export * from './invite'
