+++ src/pages/invite/components/error-fallback/error-modal/index.ts
@@ -1,0 +1,5 @@
+export {
+  InviteErrorMoal,
+  InviteErrorMoalForbidden,
+  InviteErrorMoalLogin,
+} from './ErrorModal'
