+++ src/pages/invite/components/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { InviteErrorFallback } from './ErrorFallback'
