+++ src/pages/invite/components/index.ts
@@ -1,0 +1,1 @@
+export * from './error-fallback'
