+++ src/pages/invite/index.ts
@@ -1,0 +1,1 @@
+export { default as InvitePage } from './InvitePage'
